Fix interactive window detection
