+++ vscode.proposed.d.ts
@@ -184,15 +184,6 @@ declare module 'vscode' {
 		 * Defaults to false.
 		 */
 		forceNewSession?: boolean | { detail: string };
-		/**
-		 * Whether we should show the indication to sign in in the Accounts menu.
-		 *
-		 * If false, the user will be shown a badge on the Accounts menu with an option to sign in for the extension.
-		 * If true, no indication will be shown.
-		 *
-		 * Defaults to false.
-		 */
-		silent?: boolean;
 	}
 
 	export namespace authentication {
@@ -1125,18 +1116,6 @@ declare module 'vscode' {
 		 * An optional flag to sort the final results by index of first query match in label. Defaults to true.
 		 */
 		sortByLabel: boolean;
-	}
-
-	//#endregion
-
-	//#region https://github.com/microsoft/vscode/issues/132068
-
-	export interface QuickPick<T extends QuickPickItem> extends QuickInput {
-
-		/*
-		 * An optional flag to maintain the scroll position of the quick pick when the quick pick items are updated. Defaults to false.
-		 */
-		keepScrollPosition?: boolean;
 	}
 
 	//#endregion
@@ -2812,20 +2791,6 @@ declare module 'vscode' {
 
 	//#endregion
 
-	//#region https://github.com/microsoft/vscode/issues/88716
-	export interface QuickPickItem {
-		buttons?: QuickInputButton[];
-	}
-	export interface QuickPick<T extends QuickPickItem> extends QuickInput {
-		readonly onDidTriggerItemButton: Event<QuickPickItemButtonEvent<T>>;
-	}
-	export interface QuickPickItemButtonEvent<T extends QuickPickItem> {
-		button: QuickInputButton;
-		item: T;
-	}
-
-	//#endregion
-
 	//#region @eamodio https://github.com/microsoft/vscode/issues/133935
 
 	export interface SourceControl {
@@ -2834,14 +2799,4 @@ declare module 'vscode' {
 
 	//#endregion
 
-	//#region @sandy081 https://github.com/microsoft/vscode/issues/132183
-
-	export interface OutputChannel {
-		/*
-		 * Replaces the existing contents of the channel with the given value.
-		 */
-		replace(value: string): void;
-	}
-
-	//#endregion
 }
